requirements/requirements-base.txt
django-jsonfield>=1.0.1
stripe>=1.49.0
djangorestframework>=3.6.0
simplejson>=3.10.0
